chore: migrate symlinks test file into file-system test file
