+++ test/file-system.test.ts
@@ -1,213 +1,162 @@
-import type { Dirent, Stats } from "node:fs";
-import { fs } from "memfs";
-import { afterEach, describe, expect, it, vi } from "vitest";
+import { readFileSync } from "node:fs";
+import { readFile } from "node:fs/promises";
+import { describe, expect, it } from "vitest";
 import { fromFileSystem, fromFileSystemSync } from "../src/file-system";
+import { symlink, testdir, testdirSync } from "../src/utils";
 
-vi.mock("node:fs", async () => {
-  const memfs: { fs: typeof fs } = await vi.importActual("memfs");
-
-  return memfs.fs;
-});
-
-vi.mock("node:fs/promises", async () => {
-  const memfs: { fs: typeof fs } = await vi.importActual("memfs");
-
-  return memfs.fs.promises;
-});
-
-afterEach(() => {
-  vi.clearAllMocks();
-});
-
-describe("fromFileSystem", () => {
+describe("invalid paths or directories", () => {
   it("should return an empty object if the path does not exist", async () => {
-    vi.spyOn(fs.promises, "stat").mockRejectedValueOnce(new Error("Path does not exist"));
-
     const result = await fromFileSystem("non-existent-path");
 
     expect(result).toEqual({});
   });
 
-  it("should return an empty object if the path is not a directory", async () => {
-    vi.spyOn(fs.promises, "stat").mockResolvedValueOnce({
-      isDirectory: () => false,
-    } as Stats);
-
-    const result = await fromFileSystem("not-a-directory");
-
-    expect(result).toEqual({});
-  });
-
-  it("should return the directory structure with file contents", async () => {
-    const mockFiles = {
-      "file1.txt": "content1",
-      "file2.txt": "content2",
-      "subdir": {
-        "file3.txt": "content3",
-      },
-    };
-
-    // @ts-expect-error - TODO: fix this
-    vi.spyOn(fs.promises, "stat").mockImplementation(async (path: string) => {
-      if (path === "test-dir" || path === "test-dir/subdir") {
-        return {
-          isDirectory: () => true,
-          isFile: () => false,
-          isSymbolicLink: () => false,
-        } as Stats;
-      }
-      return {
-        isDirectory: () => false,
-        isFile: () => true,
-        isSymbolicLink: () => false,
-      } as Stats;
-    });
-
-    // @ts-expect-error - TODO: fix this
-    vi.spyOn(fs.promises, "readdir").mockImplementation(async (path: string) => {
-      if (path === "test-dir") {
-        return [
-          {
-            name: "file1.txt",
-            isDirectory: () => false,
-            isFile: () => true,
-            isSymbolicLink: () => false,
-          },
-          {
-            name: "file2.txt",
-            isDirectory: () => false,
-            isFile: () => true,
-            isSymbolicLink: () => false,
-          },
-          {
-            name: "subdir",
-            isDirectory: () => true,
-            isFile: () => false,
-            isSymbolicLink: () => false,
-          },
-        ] as Dirent[];
-      }
-      if (path === "test-dir/subdir") {
-        return [
-          {
-            name: "file3.txt",
-            isDirectory: () => false,
-            isFile: () => true,
-            isSymbolicLink: () => false,
-          },
-        ] as Dirent[];
-      }
-      return [];
-    });
-
-    // @ts-expect-error - TODO: fix this
-    vi.spyOn(fs.promises, "readFile").mockImplementation(async (path: string) => {
-      if (path === "test-dir/file1.txt") return "content1";
-      if (path === "test-dir/file2.txt") return "content2";
-      if (path === "test-dir/subdir/file3.txt") return "content3";
-      return "";
-    });
-
-    const result = await fromFileSystem("test-dir");
-
-    expect(result).toMatchObject(mockFiles);
-  });
-});
-
-describe("fromFileSystemSync", () => {
-  it("should return an empty object if the path does not exist", () => {
-    vi.spyOn(fs, "statSync").mockImplementationOnce(() => {
-      throw new Error("Path does not exist");
-    });
-
+  it("should return an empty object if the path does not exist (sync)", () => {
     const result = fromFileSystemSync("non-existent-path");
 
     expect(result).toEqual({});
   });
 
-  it("should return an empty object if the path is not a directory", () => {
-    vi.spyOn(fs, "statSync").mockResolvedValueOnce({
-      isDirectory: () => false,
-      isFile: () => false,
-      isSymbolicLink: () => false,
-    } as any);
-
-    const result = fromFileSystemSync("not-a-directory");
+  it("should return an empty object if the path is not a directory", async () => {
+    const result = await fromFileSystem("not-a-directory");
 
     expect(result).toEqual({});
   });
 
-  it("should return the directory structure with file contents", () => {
+  it("should return an empty object if the path is not a directory (sync)", () => {
+    const result = fromFileSystemSync("not-a-directory");
+
+    expect(result).toEqual({});
+  });
+});
+
+describe("handle symlinks", () => {
+  it("should correctly handle symbolic links in the directory", async () => {
     const mockFiles = {
-      "file1.txt": "content1",
-      "file2.txt": "content2",
-      "subdir": {
-        "file3.txt": "content3",
+      "file1.txt": "content1\n",
+      "symlink.txt": symlink("file1.txt"),
+      "symlinked-dir": symlink("nested"),
+      "nested": {
+        "file2.txt": "content2\n",
+        "link-to-parent.txt": symlink("../file1.txt"),
+        "double-nested": {
+          "file3.txt": "content3\n",
+          "link-to-parent.txt": symlink("../../file1.txt"),
+          "double-double-nested": {
+            "README.md": symlink("../../../../../../README.md"),
+          },
+        },
       },
     };
 
-    // @ts-expect-error - TODO: fix this
-    vi.spyOn(fs, "statSync").mockImplementation((path: string) => {
-      if (path === "test-dir" || path === "test-dir/subdir") {
-        return {
-          isDirectory: () => true,
-          isFile: () => false,
-          isSymbolicLink: () => false,
-        } as Stats;
-      }
-      return {
-        isDirectory: () => false,
-        isFile: () => true,
-        isSymbolicLink: () => false,
-      } as Stats;
+    const result = await fromFileSystem("./test/fixtures/symlinks");
+
+    expect(result).toMatchObject(mockFiles);
+  });
+
+  it("should correctly handle symbolic links in the directory (sync)", () => {
+    const mockFiles = {
+      "file1.txt": "content1\n",
+      "symlink.txt": symlink("file1.txt"),
+      "symlinked-dir": symlink("nested"),
+      "nested": {
+        "file2.txt": "content2\n",
+        "link-to-parent.txt": symlink("../file1.txt"),
+        "double-nested": {
+          "file3.txt": "content3\n",
+          "link-to-parent.txt": symlink("../../file1.txt"),
+          "double-double-nested": {
+            "README.md": symlink("../../../../../../README.md"),
+          },
+        },
+      },
+    };
+
+    const result = fromFileSystemSync("./test/fixtures/symlinks");
+
+    expect(result).toMatchObject(mockFiles);
+  });
+
+  it("should handle symbolic links using testdir", async () => {
+    const files = await fromFileSystem("./test/fixtures/symlinks");
+
+    const path = await testdir(files);
+
+    const rootReadme = await readFile("./README.md", "utf8");
+    const testdirReadme = await readFile(`${path}/nested/double-nested/double-double-nested/README.md`, "utf8");
+
+    expect(rootReadme).toStrictEqual(testdirReadme);
+  });
+
+  it("should handle symbolic links using testdir (sync)", () => {
+    const files = fromFileSystemSync("./test/fixtures/symlinks");
+
+    const path = testdirSync(files);
+
+    const rootReadme = readFileSync("./README.md", "utf8");
+    const testdirReadme = readFileSync(`${path}/nested/double-nested/double-double-nested/README.md`, "utf8");
+
+    expect(rootReadme).toStrictEqual(testdirReadme);
+  });
+
+  it("should handle symbolic links using testdir with custom path", async () => {
+    const files = await fromFileSystem("./test/fixtures/symlinks");
+
+    const path = await testdir(files, {
+      dirname: "./three/levels/deep",
     });
 
-    // @ts-expect-error - TODO: fix this
-    vi.spyOn(fs, "readdirSync").mockImplementation((path: string) => {
-      if (path === "test-dir") {
-        return [
-          {
-            name: "file1.txt",
-            isDirectory: () => false,
-            isFile: () => true,
-            isSymbolicLink: () => false,
-          },
-          {
-            name: "file2.txt",
-            isDirectory: () => false,
-            isFile: () => true,
-            isSymbolicLink: () => false,
-          },
-          {
-            name: "subdir",
-            isDirectory: () => true,
-            isFile: () => false,
-            isSymbolicLink: () => false,
-          },
-        ] as Dirent[];
-      }
-      if (path === "test-dir/subdir") {
-        return [
-          {
-            name: "file3.txt",
-            isDirectory: () => false,
-            isFile: () => true,
-            isSymbolicLink: () => false,
-          },
-        ] as Dirent[];
-      }
-      return [];
+    const rootReadme = await readFile("./README.md", "utf8");
+    const testdirReadme = await readFile(`${path}/nested/double-nested/double-double-nested/README.md`, "utf8");
+
+    expect(rootReadme).toStrictEqual(testdirReadme);
+  });
+
+  it("should handle symbolic links using testdir with custom path (sync)", () => {
+    const files = fromFileSystemSync("./test/fixtures/symlinks");
+
+    const path = testdirSync(files, {
+      dirname: "./three/levels/deep-sync",
     });
 
-    // @ts-expect-error - TODO: fix this
-    vi.spyOn(fs, "readFileSync").mockImplementation((path: string) => {
-      if (path === "test-dir/file1.txt") return "content1";
-      if (path === "test-dir/file2.txt") return "content2";
-      if (path === "test-dir/subdir/file3.txt") return "content3";
-      return "";
-    });
+    const rootReadme = readFileSync("./README.md", "utf8");
+    const testdirReadme = readFileSync(`${path}/nested/double-nested/double-double-nested/README.md`, "utf8");
 
-    const result = fromFileSystemSync("test-dir");
+    expect(rootReadme).toStrictEqual(testdirReadme);
+  });
+});
+
+describe("map file contents", () => {
+  it("should return the directory structure with file contents", async () => {
+    const mockFiles = {
+      "file.txt": "this is just a file!\n",
+      "README.md": "# vitest-testdirs\n",
+      "nested": {
+        "README.md": "# Nested Fixture Folder\n",
+        // TODO: use buffer after https://github.com/luxass/vitest-testdirs/issues/66 is fixed
+        // "image.txt": Buffer.from([72, 101, 108, 108, 111, 44, 32, 87, 111, 114, 108, 100, 33, 10]),
+        "image.txt": "Hello, World!\n",
+      },
+    };
+
+    const result = await fromFileSystem("./test/fixtures/file-system/test-dir");
+
+    expect(result).toMatchObject(mockFiles);
+  });
+
+  it("should return the directory structure with file contents (sync)", () => {
+    const mockFiles = {
+      "file.txt": "this is just a file!\n",
+      "README.md": "# vitest-testdirs\n",
+      "nested": {
+        "README.md": "# Nested Fixture Folder\n",
+        // TODO: use buffer after https://github.com/luxass/vitest-testdirs/issues/66 is fixed
+        // "image.txt": Buffer.from([72, 101, 108, 108, 111, 44, 32, 87, 111, 114, 108, 100, 33, 10]),
+        "image.txt": "Hello, World!\n",
+      },
+    };
+
+    const result = fromFileSystemSync("./test/fixtures/file-system/test-dir");
 
     expect(result).toMatchObject(mockFiles);
   });
